refactor(web): split terminal panel state helpers
